+++ src/Machina/EventArgs/ActionReleasedArgs.cs
@@ -62,7 +62,7 @@ namespace Machina
 
         public override string ToJSONString()
         {
-            return string.Format("{{\"event\":\"action-released\",\"last\":\"{0}\",\"pend\":{1},\"pos\":{2},\"ori\":{3},\"quat\":{4},\"axes\":{5},\"extax\":{6},\"conf\":{7}}}",
+            return string.Format("{{\"event\":\"action-released\",\"last\":\"{0}\",\"id\":\"{8}\",\"pend\":{1},\"pos\":{2},\"ori\":{3},\"quat\":{4},\"axes\":{5},\"extax\":{6},\"conf\":{7}}}",
                 Util.EscapeDoubleQuotes(this.LastAction.ToInstruction()),
                 this.PendingReleaseToDevice,
                 this.Position?.ToArrayString() ?? "null",
@@ -70,7 +70,8 @@ namespace Machina
                 this.Rotation?.Q.ToArrayString() ?? "null",
                 this.Axes?.ToArrayString() ?? "null",
                 this.ExternalAxes?.ToArrayString() ?? "null",
-                "null");  // placeholder for whenever IK are introduced...
+                "null",  // placeholder for whenever IK solvers are introduced...
+                this.LastAction.Id);
         }
     }