+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Rename main `Control` `RobotCursor`s to match the issue/release/execute pattern.
 - New `ActionExecuted` event, combines former `MotionCursorUpdated` and `ActionCompleted` in the same event.
 - New `ActionReleased` event, raised any time an `Action` is sent to the device for execution.
+- New `ActionIssued` event, raised any time an `Action` is successfully issued and scheduled for release or compilation. 
 
 
 ## BUILD 1415